서치와 디테일 페이지 만드는법 수정
search?keyword= , detailSeriese?id= 요청으로 가져오게 수정
